fixed: 修复ts打包脚本遗漏部分导出组件的问题
